Extract error module
